fix: post-audit residual cleanup (8 user stories)
Split oversized files to comply with 500 NLOC limit:
- pq.rs (1828→486) split into pq_kmeans.rs, pq_opq.rs, pq_tests.rs
- rabitq.rs (979→452) tests extracted to rabitq_tests.rs
- inverted_index.rs (928→630) tests extracted to inverted_index_tests.rs
- persistence.rs (844→574) tests extracted to persistence_tests.rs
- conditions.rs (613→399) vector parsing extracted to condition_vectors.rs

Refactored functions exceeding 50 NLOC:
- lifecycle.rs: extracted init_collection_parts/assemble/rebuild_bm25_index helpers
- search.rs: maxscore_search reduced from 75 to 39 NLOC

Implemented UNION/INTERSECT/EXCEPT execution in VelesQL query engine
(previously parsed but silently ignored). Added set_operations.rs with
7 end-to-end tests. Updated conformance matrix.
